Separate model into sb3 / rllib implementations, add install script
