+++ experiments/sb3.py
@@ -7,8 +7,7 @@ import psutil
 import sumo_rl
 import supersuit as ss
 import torch
-from array2gif import write_gif
-from custom.model import CustomActorCriticPolicy
+from custom.sb3_model import CustomActorCriticPolicy
 from custom.utils import load_cfg, smooth_data
 from stable_baselines3 import PPO, SAC
 from stable_baselines3.common.callbacks import BaseCallback
